Only match clearing number with regex
Enable assisting a user in entering their account number by as early
as possible matching the clearing number to a bank. This allows for
more granular error message, rather than just saying that the account
number is incorrect until every digit is entered correctly.

That there are only digits provided is ensured by the regex replace of
non-digit characters, while that the right number of digits are provided
is check by the subsequent length check.