Group account number formats as in Bankgirot docs
Rather than repeating the lengths configuration for each bank, opening
up for mistakes, group the different types of account number lengths
in a constant that directly maps to the two main types of account
numbers and their comments in the Bankgirot documentation:
https://www.bankgirot.se/globalassets/dokument/anvandarmanualer/bankernaskontonummeruppbyggnad_anvandarmanual_sv.pdf

This results in 5 distinct sets of clearing number, account number &
control length combinations.

Definition of length keys:
- clearing: length of clearing number
- account: length of account number without clearing number
- control: number of last digits in whole account number
  (`clearing + account`) for mod 10/11 checksum calculation

This means the total account number length is `clearing + account` of
which the last `control` number of digits are used for the mod 10/11
checksum calculation.

AB-3716